Add use-search-bar
